+++ djconfig/__init__.py
@@ -3,7 +3,7 @@
 from __future__ import unicode_literals
 
 from . import conf
-
+from . import admin
 
 default_app_config = 'djconfig.apps.DjConfig'
 
@@ -12,4 +12,4 @@ register = conf.register
 reload_maybe = conf.reload_maybe
 
 __version__ = "0.8.0"
-__all__ = ['config', 'register', 'reload_maybe']
+__all__ = ['config', 'register', 'reload_maybe', 'admin']
